+++ examples/ast.rs
@@ -7,7 +7,13 @@ fn main() {
     }
 
     let text = std::fs::read_to_string(&args[1]).unwrap();
-    match roxmltree::Document::parse(&text) {
+
+    // Allow DTD for this example.
+    let opt = roxmltree::ParsingOptions {
+        allow_dtd: true,
+    };
+
+    match roxmltree::Document::parse_with_options(&text, opt) {
         Ok(doc) => print!("{:?}", doc),
         Err(e) => println!("Error: {}.", e),
     }