+++ tests/ast.rs
@@ -55,8 +55,12 @@ fn compare_ast() {
 fn actual_test(path: path::PathBuf) {
     let expected = load_file(&path.with_extension("yaml"));
 
+    let opt = ParsingOptions {
+        allow_dtd: true,
+    };
+
     let input_xml = load_file(&path);
-    let doc = match Document::parse(&input_xml) {
+    let doc = match Document::parse_with_options(&input_xml, opt) {
         Ok(v) => v,
         Err(e) => {
             assert_eq!(TStr(&format!("error: \"{}\"", e)), TStr(expected.trim()));